hash: clean up comments, audit for bugs
We stopped supporting shared hash tables a long time ago.  Remove comments
describing the behavior of shared hash tables.

Add an event counter for pushing a hash to the front when it is already at
the front.

Audited the code for a bug related to bucket handling that impairs space
efficiency when the bucket size is greater than 1.  Didn't find one.

Signed-off-by: Zygo Blaxell <bees@furryterror.org>